# Conflicts:
#   backend/raonzena/src/main/java/com/ssafy/raonzena/api/controller/ProfileController.java

+++ backend/raonzena/src/main/java/com/ssafy/raonzena/api/service/LiveService.java
@@ -13,7 +13,7 @@ public interface LiveService {
     // 현재 실행중인 방 조회
     List<LiveRoomInfoRes> findRooms(Map<String, Object> conditions);
     // 팔로잉 유저들의 방 조회
-    List<LiveRoomInfoRes> findFollowingRooms(int sessionUserNo); /////////세션정보 필요//////////
+    List<LiveRoomInfoRes> findFollowingRooms(long sessionUserNo); /////////세션정보 필요//////////
     // 유저 게임 참가 가능 여부 조회
-    boolean isAccessible(int roomNo, int sessionHeadCount); /////////세션정보 필요//////////
+    boolean isAccessible(long roomNo, int sessionHeadCount); /////////세션정보 필요//////////
 }
